hw7 done

+++ src/main/kotlin/ru/tinkoff/fintech/homework/lesson6/db/JdbcStorageDao.kt
@@ -1,6 +1,5 @@
 package ru.tinkoff.fintech.homework.lesson6.db
 
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Profile
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.stereotype.Service
@@ -13,31 +12,29 @@ class JdbcStorageDao(private val jdbcTemplate: JdbcTemplate) : StorageDao {
     override fun getConfig(id: Int): Config {
         val response = jdbcTemplate.query("select * from configs where id = $id limit 1") { rs, _ ->
             Config(
-                rs.getLong("id"),
+                rs.getInt("id"),
                 rs.getInt("diskSize"),
                 rs.getInt("cores"),
                 rs.getInt("sizeRam"),
             )
         }
-        if (response.size != 0){
+        if (response.size != 0) {
             return response[0]
-        }
-        else throw IllegalArgumentException("No config with this id")
+        } else throw IllegalArgumentException("No config with this id")
     }
 
     override fun getImg(id: Int): Image {
-        val response =  jdbcTemplate.query("select * from images where id = $id limit 1") { rs, _ ->
+        val response = jdbcTemplate.query("select * from images where id = $id limit 1") { rs, _ ->
             Image(
-                rs.getLong("id"),
+                rs.getInt("id"),
                 rs.getString("name"),
                 rs.getString("type"),
                 rs.getInt("diskSizeRequirements"),
                 rs.getInt("ramRequirements"),
             )
         }
-        if (response.size != 0){
+        if (response.size != 0) {
             return response[0]
-        }
-        else throw IllegalArgumentException("No image with this id")
+        } else throw IllegalArgumentException("No image with this id")
     }
 }
