+++ src/main/kotlin/ru/tinkoff/fintech/homework/lesson6/vm/model/Vm.kt
@@ -5,8 +5,8 @@ import ru.tinkoff.fintech.homework.lesson6.vm.model.external.VmStatus
 
 interface Vm {
     val type: String
-    val imageId: Long?
-    val configId: Long?
+    val imageId: Int?
+    val configId: Int?
     val osType: String
     val state: VmState
     val vmStatus: VmStatus
